Refactor to support more custom formats
