+++ packages/ember-repl/addon/src/browser/compile/formats/gjs/babel.ts
@@ -1,0 +1,8 @@
+import type { availablePlugins, availablePresets, transform } from '@babel/standalone';
+
+export interface Babel {
+  availablePlugins: typeof availablePlugins;
+  availablePresets: typeof availablePresets;
+  transform: typeof transform;
+}
+
