+++ packages/ember-repl/addon/src/browser/compile/formats/gjs/eval.ts
@@ -1,12 +1,11 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { modules } from '../known-modules.ts';
+import { modules } from './known-modules.ts';
 
-import type { ExtraModules } from '../types.ts';
 import type Component from '@glimmer/component';
 
 export function evalSnippet(
   compiled: string,
-  extraModules: ExtraModules = {}
+  extraModules: Record<string, unknown /* imported module */> = {}
 ): {
   default: Component;
   services?: { [key: string]: unknown };
