+++ packages/ember-repl/addon/src/browser/compile/formats/gjs/known-modules.ts
