+++ packages/ember-repl/addon/src/browser/compile/types.ts
@@ -1,4 +1,5 @@
-import { type Plugin } from 'unified';
+import type { ComponentLike } from '@glint/template';
+import type { Plugin } from 'unified';
 
 export interface EvalImportMap {
   [moduleName: string]: ScopeMap;
@@ -9,3 +10,19 @@ export interface ScopeMap {
 }
 
 export type UnifiedPlugin = Plugin; // Parameters<ReturnType<typeof unified>['use']>[0];
+
+
+export interface CompileResult {
+  component?: ComponentLike;
+  error?: Error;
+  name: string;
+}
+
+export type Options = {
+  /**
+   * @internal
+   * @deprecated do not use - not under semver
+   */
+  skypack?: boolean;
+};
+
