+++ packages/ember-repl/addon/src/browser/index.ts
@@ -1,8 +1,5 @@
 export { compile, Compiled } from './compile/index.ts';
-export { compileHBS } from './hbs.ts';
-export { compileJS } from './js.ts';
-export { modules as frameworkModules } from './known-modules.ts';
-export { invocationName, invocationOf, nameFor } from './utils.ts';
+export { invocationName, invocationOf, nameFor } from './compile/utils.ts';
 
 // Public Types
-export type { CompileResult } from './types';
+export type { CompileResult } from './compile/types';
